server/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
gunicorn
python-dotenv